% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gRodon.R
\name{fitGCModels}
\alias{fitGCModels}
\title{Fit gRodon GC-corrected models}
\usage{
fitGCModels(stat_data, stat_data_extremo)
}
\arguments{
\item{stat_data}{dataframe with codon usage statistics and known doubling times}
}
\description{
This function fits the gRodon GC-corrected metagenome mode models
}
