% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_copiotrophy.R
\name{fitIoC}
\alias{fitIoC}
\title{Create an Index of Copiotrophy}
\usage{
fitIoC(x = NULL, dCUB = NA, nCAZy = NA, nGenes = NA, completeness = 1)
}
\arguments{
\item{x}{Dataframe with the columns "dCUB", "nCAZy", and "nGenes" containing dCUB
values from gRodon's predictGrowth() function, the number of CAZymes per genome
(as predicted by dbCAN or other tool), and the number of protein coding genes per genome
respectively. If x is omitted, these values can be supplied as vectors.}

\item{dCUB}{Vector of dCUB values from gRodon's predictGrowth() function. Only required
if x is not provided as an argument.}

\item{nCAZy}{Vector of the number of CAZymes per genome. Only required
if x is not provided as an argument.}

\item{nGenes}{Vector of the number of protein coding genes per genome. Only required
if x is not provided as an argument.}

\item{completeness}{Vector of completeness values (\%) for each genome (e.g., from checkM2).
Can be useful when dealing with partial MAGs.}
}
\value{
fitIoC returns a list with the following elements:
\describe{
\item{IoC}{The calculated IoC values for each of the genomes in x}
\item{model}{The fitted IoC model that can be passed to the predictIoC() function
to place new genomes on this fitted index of copiotrophy.}
}
}
\description{
This function creates an index of copiotrophy from a set of genomes that captures
the range of growth strategies encoded in that set.
(as in https://doi.org/10.1101/2025.09.01.673550 and https://doi.org/10.1126/science.ado5323).
}
\examples{
#First let's take a look at the "mags" dataframe provided in this package
# (from https://doi.org/10.1101/2025.09.01.673550)
load(system.file('extdata',
  'mags_df.rda',
  package = 'gRodon'))
head(mags)

#Let's hold out 3 mags for illustration
mags_held_out <- mags[1:3,]
mags_train <- mags[4:nrow(mags),]

#Now let's fit an index of copiotrophy using these mags
IoC <- fitIoC(mags_train)

#Now let's place our held out mags onto this fitted IoC
predictIoC(mags_held_out,model=IoC)

}
