% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gRodon.R
\name{fitModels}
\alias{fitModels}
\title{Fit gRodon models}
\usage{
fitModels(stat_data, stat_data_extremo)
}
\arguments{
\item{stat_data}{dataframe with codon usage statistics and known doubling times}
}
\description{
This function fits the gRodon models
}
