% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gRodon.R
\name{getStatistics}
\alias{getStatistics}
\title{Get Codon Statistics}
\usage{
getStatistics(
  gene_file,
  genetic_code = "11",
  bg = "all",
  trimlen = NA,
  trimside = "start"
)
}
\arguments{
\item{gene_file}{path to CDS-containing fasta file}
}
\description{
Wrapper to estimate codon statistics  for an organism (used to fit gRodon).
It is assumed that gene names contain annotations of ribosomal proteins.
}
