% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gRodon.R
\name{getStatisticsBatch}
\alias{getStatisticsBatch}
\title{Get Codon Statistics for All Genomes In a Directory}
\usage{
getStatisticsBatch(directory, mc.cores = 1)
}
\arguments{
\item{directory}{path to directory containing annotated CDS files}
}
\description{
This function gets the codon usage statistics for all CDS files in a directory (used to fit gRodon).
It is assumed that gene names contain annotations of ribosomal proteins.
}
