% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_growth.R
\name{predictGrowth}
\alias{predictGrowth}
\title{Predict Growth Rate}
\usage{
predictGrowth(
  genes,
  highly_expressed,
  metagenome = FALSE,
  temperature = "none"
)
}
\arguments{
\item{genes}{DNAStringSet object holding all nucleotide sequences from a
genome.}

\item{highly_expressed}{Logical vector describing the set of highly expressed
genes. Must be of same lend as \code{genes}.}

\item{metagenome}{Whether to run prediction in metagenome mode (no codon pair
bias)}

\item{temperature}{Optimal growth temperature. By default this is set as
"none" and we do not guarantee good results for non-mesophilic organisms.}
}
\value{
A list with the following components is returned:
\describe{
\item{CUBHE}{Median codon usage bias of the highly expressed genes (MILC)
calculated using the genome-wide codon usage as the expected bias}
\item{ConsistencyHE}{Median codon usage bias of the highly expressed genes
(MILC) calculated using the codon usage of highly expressed genes as the
expected bias}
\item{CPB}{Genome-wide codon pair bias (Coleman et al. 2008)}
\item{FilteredSequences}{Number of gene sequences filtered out during
calulation (due to length and/or presence of ambiguous bases)}
\item{d}{Predicted doubling time}
\item{LowerCI}{Lower CI of \code{d} (2.5\%)}
\item{UpperCI}{Upper CI of \code{d} (97.5\%)}
}
}
\description{
This function predicts the growth rate of a prokaryotic organism based on
genomic codon usage patterns (Weissman et. al. TBD).
}
\examples{
genes <- readDNAStringSet(path_to_genome)
highly_expressed <- grepl("ribosomal protein",names(genes),ignore.case = T)
getGrowth(genes, highly_expressed)
}
