% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_growth.R
\name{predictGrowth}
\alias{predictGrowth}
\title{Predict Growth Rate}
\usage{
predictGrowth(
  genes,
  highly_expressed,
  mode = "full",
  temperature = "none",
  training_set = "madin",
  depth_of_coverage = NULL,
  fragments = FALSE,
  genetic_code = NULL,
  n_le = 100,
  bg = "all"
)
}
\arguments{
\item{genes}{DNAStringSet object holding all in-frame coding sequences (nucleotide sequences)
from a genome. See Biostrings package.}

\item{highly_expressed}{Logical vector describing the set of highly expressed
genes. Must be of same length as \code{genes}. Typically these are ribosomal proteins
(all models were trained using ribosomal proteins as the highly expressed set.)}

\item{mode}{Whether to run prediction in full, partial, metagenome_v1, metagenome_v2,
metagenome_150bp, metagenome_250bp,
eukaryote, or metagenome_euk mode
(by default gRodon applies the full model). Mode metagenome_v2 may run slower
than the other prediction modes when consistency>0.6.}

\item{temperature}{Optimal growth temperature. By default this is set as
"none" for prokaryotes and we do not guarantee good results for non-mesophilic
since prokaryotes few were used to fit the model. For eukaryotes, though, including
an optimal growth temperature drastically improves model predictions as there is
a very strong relationship between OGT and max. growth rate for these organisms.}

\item{training_set}{Whether to use models trained on the original Vieira-Silva et al.
doubling time dataset or doubling times drawn from the Madin et al. database. This
setting is only used for prokaryotic modes (eukaryotic models based on their own
training set from Weissman et al. TBD). For metagenome_v2,metagenome_150bp, and metagenome_250bp, modes, only the madin
set is available. By default training set is now set to Madin. For AOA and NOB try AOA_NOB
which includes an expanded set of these organisms (including some measurements from enrichment cultures)}

\item{depth_of_coverage}{When using metagenome mode, provide a vector containing
the coverage of your ORFs to improve your estimate}

\item{fragments}{Do not change (performance will suffer, publication forthcoming).
If using gene fragments predicted from reads, will use a
more permissive length filter (120bp as opposed to 240bp)}

\item{genetic_code}{The genetic code of the organism to be used in codon usage
calculations (see https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi). By
default code "1" is used for eukaryote mode and code "11" for prokaryotic modes.}

\item{n_le}{The number of genes to sample to generate a background estimate of
codon usage against which to calculate the codon usage of the highly expressed
genes. Only used by metagenome_v2 mode and only when consistency>0.6. Increasing
will make prediction slower, but may decrease variance. Decreasing is not recommended
but will increase speed.}

\item{bg}{Feature in testing, do not change}
}
\value{
gRodon returns a list with the following elements:
\describe{
\item{CUBHE}{Median codon usage bias of the highly expressed genes (MILC)
calculated using the genome-wide codon usage as the expected bias}
\item{ConsistencyHE}{Mean codon usage bias of the highly expressed genes
(MILC) calculated using the codon usage of highly expressed genes as the
expected bias}
\item{CPB}{Genome-wide codon pair bias (Coleman et al. 2008)}
\item{FilteredSequences}{Number of gene sequences filtered out during
calulation (due to length and/or presence of ambiguous bases)}
\item{d}{Predicted doubling time in hours}
\item{LowerCI}{Lower CI of \code{d} (2.5\%) from linear model}
\item{UpperCI}{Upper CI of \code{d} (97.5\%) from linear model}
\item{MILC}{MILC of highly expressed genes (only short-read modes return)}
\item{ENCprime}{ENCprime of highly expressed genes (only short-read modes return)}
\item{B}{B of highly expressed genes (only short-read modes return)}
\item{SCUO}{SCUO of highly expressed genes (only short-read modes return)}
\item{MCB}{MCB of highly expressed genes (only short-read modes return)}
}
}
\description{
This function predicts the growth rate of prokaryotic or eukaryotic organisms based on
genomic codon usage patterns (Weissman et. al. 2021; Weissman et al. TBD).
}
\examples{
# Load in example genome (Streptococcus pyogenes M1, downloaded from RefSeq)
# included with gRodon
library(Biostrings)
path_to_genome <- system.file('extdata',
  'GCF_000349925.2_ASM34992v2_cds_from_genomic.fna.gz',
  package = 'gRodon')
genes <- readDNAStringSet(path_to_genome)

# Search pre-existing annotations for ribosomal proteins, which we
# will use as our set of highly expressed genes
highly_expressed <- grepl("^(?!.*(methyl|hydroxy)).*0S ribosomal protein",names(genes),ignore.case = TRUE, perl = TRUE)

# Run the gRodon growth prediction pipeline
predictGrowth(genes, highly_expressed)

# Run gRodon with temperature option (not needed for mesophiles, gRodon not
# validated on extremophiles, use with care)
predictGrowth(genes, highly_expressed, temperature = 37)

}
