% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_copiotrophy.R
\name{predictIoC}
\alias{predictIoC}
\title{Calculate Index of Copiotrophy for a Set of Genomes}
\usage{
predictIoC(x = NULL, dCUB = NA, nCAZy = NA, nGenes = NA, model)
}
\arguments{
\item{x}{Dataframe with the columns "dCUB", "nCAZy", and "nGenes" containing dCUB
values from gRodon's predictGrowth() function, the number of CAZymes per genome
(as predicted by dbCAN or other tool), and the number of protein coding genes per genome
respectively. If x is omitted, these values can be supplied as vectors.}

\item{dCUB}{Vector of dCUB values from gRodon's predictGrowth() function. Only required
if x is not provided as an argument.}

\item{nCAZy}{Vector of the number of CAZymes per genome. Only required
if x is not provided as an argument.}

\item{nGenes}{Vector of the number of protein coding genes per genome. Only required
if x is not provided as an argument.}

\item{model}{The output of fitIoC(), which creates an index of copiotrophy from a set of genomes that captures
the range of growth strategies encoded in that set. Alternatively, set to "permafrost" to use the IoC
computed in https://doi.org/10.1101/2025.09.01.673550 from MAGs from a permafrost warming experiment
or "pacific" to use the IoC from the P16 Pacific ocean transect in https://doi.org/10.1126/science.ado5323}
}
\value{
predictIoC returns a vector of IoC values
}
\description{
This function calculates IoC values, given some already fitted index of copiotrophy
(as in https://doi.org/10.1101/2025.09.01.673550 and https://doi.org/10.1126/science.ado5323).
}
\examples{
#First let's take a look at the "mags" dataframe provided in this package
# (from https://doi.org/10.1101/2025.09.01.673550)
load(system.file('extdata',
  'mags_df.rda',
  package = 'gRodon'))
head(mags)

#Let's hold out 3 mags for illustration
mags_held_out <- mags[1:3,]
mags_train <- mags[4:nrow(mags),]

#Now let's fit an index of copiotrophy using these mags
IoC <- fitIoC(mags_train)

#Now let's place our held out mags onto this fitted IoC
predictIoC(mags_held_out,model=IoC)

#Alternatively, use one of the pre-computed IoC models available in this package
predictIoC(mags_held_out,model="permafrost")
predictIoC(mags_held_out,model="pacific")

}
